fixed special blocks

+++ TableProject/Assets/Scripts/Net/Shared/Net_FoundRiddleMsg.cs
@@ -25,7 +25,7 @@ public class Net_FoundRiddleMsg : NetMessage
     }
     public override void Serialize(ref DataStreamWriter writer)
     {
-        writer.WriteByte((byte)Code);
+        writer.WriteInt((int)Code);
         writer.WriteFixedString128(RiddleMessage);
     }
     public override void Deserialize(DataStreamReader reader)
