+++ TableProject/Assets/Scripts/Net/Shared/Net_RiddleAnswerMsg.cs
@@ -32,7 +32,7 @@ public class Net_RiddleAnswerMsg : NetMessage
     }
     public override void Serialize(ref DataStreamWriter writer)
     {
-        writer.WriteByte((byte)Code);
+        writer.WriteInt((int)Code);
         writer.WriteFixedString128(RiddleMessage);
         writer.WriteInt(isCorrectAnswer);
     }
